Use uint32 atomics
int64 atomics are not supported on 32-bit systems as of now: https://github.com/golang/go/issues/599

+++ traversal/stats.go
@@ -6,9 +6,9 @@ import (
 
 type Stats struct {
 	// Count of (probably) distinct addresses we've sent traversal queries to. Accessed with atomic.
-	NumAddrsTried int64
+	NumAddrsTried uint32
 	// Number of responses we received to queries related to this traversal. Accessed with atomic.
-	NumResponses int64
+	NumResponses uint32
 }
 
 func (me Stats) String() string {
